backend/requirements.txt
uvicorn
fastapi
openai
supabase
python-dotenv
httpx
python-multipart
